requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
httpx
pytest
pytest-asyncio
python-dateutil
